+++ sha256.h
@@ -30,5 +30,5 @@ typedef struct {
 void sha256_init(SHA256_CTX *ctx);
 void sha256_update(SHA256_CTX *ctx, const BYTE data[], int len);
 void sha256_final(SHA256_CTX *ctx, BYTE hash[]);
-void sha256_main(BYTE text[],int size_of_text,BYTE buf[SHA256_BLOCK_SIZE]); // 传入字符数组  以及接受数组
+void sha256_main(BYTE text[],size_t size_of_text,BYTE buf[SHA256_BLOCK_SIZE]); // 传入字符数组  以及接受数组
 #endif   // SHA256_H
